requirements.txt
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
fastapi==0.54.2
h11==0.9.0
httptools==0.1.1
passlib==1.7.2
pkg-resources==0.0.0
psycopg2==2.8.5
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
python-dotenv==0.13.0
python-multipart==0.0.5
six==1.14.0
SQLAlchemy==1.3.17
starlette==0.13.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
